@@ -13,16 +13,17 @@ import (
 	"go.mongodb.org/mongo-driver/mongo/options"
 )
 
-func FetchAppUseCase(ctx any, appID string, deviceID *string, ip string) (*entities.Application, error) {
+func FetchAppUseCase(ctx any, appID string, deviceID string, ip string) (*entities.Application, error) {
 	appRepo := repository.ApplicationRepo()
-	app, err := appRepo.FindOneByFilter(map[string]interface{}{
-		"appID": appID,
-	}, options.FindOne().SetProjection(map[string]any{
+	app, err := appRepo.FindByID(appID, options.FindOne().SetProjection(map[string]any{
 		"name":                  1,
 		"requiredVerifications": 1,
 		"requestedFields":       1,
 		"localeRestriction":     1,
 		"description":           1,
+		"appImg":                1,
+		"appIssuer":             1,
+		"appSigningKey":         1,
 	}))
 	if err != nil {
 		logger.Error("an error occured while fethcing app details", logger.LoggerOptions{
