@@ -2,9 +2,11 @@ package controller
 
 import (
 	"context"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"net/http"
+	"reflect"
 	"time"
 
 	apperrors "authone.usepolymer.co/application/appErrors"
@@ -15,6 +17,7 @@ import (
 	application_usecase "authone.usepolymer.co/application/usecases/application"
 	"authone.usepolymer.co/application/utils"
 	"authone.usepolymer.co/entities"
+	"authone.usepolymer.co/infrastructure/auth"
 	"authone.usepolymer.co/infrastructure/cryptography"
 	"authone.usepolymer.co/infrastructure/logger"
 	server_response "authone.usepolymer.co/infrastructure/serverResponse"
@@ -46,13 +49,14 @@ func CreateApplication(ctx *interfaces.ApplicationContext[dto.ApplicationDTO]) {
 			}
 		}
 	}
-	app, apiKey := application_usecase.CreateApplicationUseCase(ctx.Ctx, ctx.Body, ctx.DeviceID, ctx.GetStringContextData("UserID"), ctx.GetStringContextData("WorkspaceID"), ctx.GetStringContextData("Email"))
+	app, apiKey, appSigningKey := application_usecase.CreateApplicationUseCase(ctx.Ctx, ctx.Body, ctx.DeviceID, ctx.GetStringContextData("UserID"), ctx.GetStringContextData("WorkspaceID"), ctx.GetStringContextData("Email"))
 	if app == nil {
 		return
 	}
 	server_response.Responder.Respond(ctx.Ctx, http.StatusCreated, "app created", map[string]any{
-		"app":    app,
-		"apiKey": apiKey,
+		"app":           app,
+		"apiKey":        apiKey,
+		"appSigningKey": appSigningKey,
 	}, nil, nil, nil, nil)
 }
 
@@ -147,18 +151,111 @@ func RefreshAppAPIKey(ctx *interfaces.ApplicationContext[any]) {
 }
 
 func ApplicationSignUp(ctx *interfaces.ApplicationContext[dto.ApplicationSignUpDTO]) {
-	_, err := application_usecase.FetchAppUseCase(ctx.Ctx, ctx.Body.AppID, ctx.DeviceID, ctx.Keys["ip"].(string))
+	app, err := application_usecase.FetchAppUseCase(ctx.Ctx, ctx.Body.AppID, ctx.DeviceID, ctx.Keys["ip"].(string))
 	if err != nil {
 		return
 	}
 	appUserRepo := repository.AppUserRepo()
-	context, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-	appUserRepo.CreateOne(context, entities.AppUser{
-		AppID:  ctx.Body.AppID,
-		UserID: ctx.GetStringContextData("UserID"),
+	appUserExists, _ := appUserRepo.CountDocs(map[string]interface{}{
+		"userID": ctx.GetStringContextData("UserID"),
+		"appID":  ctx.Body.AppID,
 	})
-	server_response.Responder.Respond(ctx.Ctx, http.StatusOK, "application signup", nil, nil, nil, nil, nil)
+	if appUserExists != 0 {
+		apperrors.ClientError(ctx.Ctx, "Seems you have already signed up for this app", nil, nil)
+		return
+	}
+	userRepo := repository.UserRepo()
+	user, _ := userRepo.FindByID(ctx.GetStringContextData("UserID"))
+	if user == nil {
+		apperrors.NotFoundError(ctx.Ctx, "This user was not found")
+		return
+	}
+	eligible := true
+	outstandingIDs := []string{}
+	for _, id := range *app.RequiredVerifications {
+		if id == "nin" {
+			if user.NIN == nil {
+				outstandingIDs = append(outstandingIDs, "nin")
+				eligible = false
+			}
+		} else {
+			if user.BVN == nil {
+				outstandingIDs = append(outstandingIDs, "bvn")
+				eligible = false
+			}
+		}
+	}
+	var results []string
+	requestedFields := map[string]any{}
+	userValue := reflect.ValueOf(*user)
+
+	for _, field := range app.RequestedFields {
+		userField := userValue.FieldByName(field.Name)
+		if !userField.IsValid() {
+			results = append(results, field.Name)
+			eligible = false
+			continue
+		}
+		var userFieldData entities.KYCData[any]
+		actualValue := userField.Interface()
+		jsonBytes, _ := json.Marshal(actualValue)
+		json.Unmarshal(jsonBytes, &userFieldData)
+
+		// If Verified field doesn't exist or is not true, add to results
+		if userFieldData.Value == nil || !userFieldData.Verified {
+			results = append(results, field.Name)
+			eligible = false
+		}
+		requestedFields[field.Name] = userFieldData.Value
+	}
+
+	payload := map[string]any{}
+	var msg string
+	if eligible {
+		msg = "Sign up successful"
+	} else {
+		msg = "Additional info is required to sign up to this app"
+		payload["missingIDs"] = outstandingIDs
+		payload["unverifiedFields"] = results
+	}
+	if eligible {
+		appUserRepo.CreateOne(context.TODO(), entities.AppUser{
+			AppID:  ctx.Body.AppID,
+			UserID: ctx.GetStringContextData("UserID"),
+		})
+		decryptedAppSigningKey, _ := cryptography.DecryptData(app.AppSigningKey, nil)
+		token, _ := auth.GenerateAppUserToken(auth.ClaimsData{
+			IssuedAt:  time.Now().Unix(),
+			ExpiresAt: time.Now().Add(time.Minute * 30).Unix(),
+			Payload:   requestedFields,
+		}, string(decryptedAppSigningKey))
+		payload["token"] = token
+		payload["expiresAt"] = time.Now().Add(time.Minute * 30)
+	}
+	server_response.Responder.Respond(ctx.Ctx, http.StatusOK, msg, payload, nil, nil, nil, nil)
+}
+
+func FetchAppUsers(ctx *interfaces.ApplicationContext[dto.FetchAppUsersDTO]) {
+	filter := map[string]interface{}{
+		"appID": ctx.Body.AppID,
+	}
+	if ctx.Body.Blocked != nil {
+		filter["blocked"] = ctx.Body.Blocked
+	}
+	if ctx.Body.Deleted != nil {
+		filter["deletedAt"] = map[string]any{"$ne": nil}
+	}
+	appUserRepo := repository.AppUserRepo()
+	users, err := appUserRepo.FindManyPaginated(filter, ctx.Body.PageSize, ctx.Body.LastID, int(ctx.Body.Sort))
+	if err != nil {
+		logger.Error("an error occured while fetching apps users", logger.LoggerOptions{
+			Key:  "appID",
+			Data: ctx.Body.AppID,
+		})
+		apperrors.UnknownError(ctx.Ctx, err)
+		return
+	}
+	server_response.Responder.Respond(ctx.Ctx, http.StatusOK, "users fetched", users, nil, nil, nil, nil)
 }
 
 func FetchUserApps(ctx *interfaces.ApplicationContext[any]) {
