@@ -7,9 +7,6 @@ import (
 
 	apperrors "authone.usepolymer.co/application/appErrors"
 	"authone.usepolymer.co/application/interfaces"
-	"authone.usepolymer.co/application/utils"
-	"authone.usepolymer.co/infrastructure/ipresolver"
-	"authone.usepolymer.co/infrastructure/logger"
 	"authone.usepolymer.co/infrastructure/useragent"
 )
 
@@ -48,39 +45,13 @@ func UserAgentMiddleware(ctx *interfaces.ApplicationContext[any], minAppVersion 
 		apperrors.UnsupportedUserAgent(ctx.Ctx)
 		return nil, false
 	}
-
-	ipLookupRes, err := ipresolver.IPResolverInstance.LookUp(clientIP)
-	if err != nil {
-		logger.Error("error looking up ip", logger.LoggerOptions{
-			Key:  "error",
-			Data: err,
-		},
-			logger.LoggerOptions{
-				Key:  "ip",
-				Data: clientIP,
-			}, logger.LoggerOptions{
-				Key:  "user agent",
-				Data: agent,
-			})
-		return nil, false
-	}
-	logger.Info("request-ip-lookup", logger.LoggerOptions{
-		Key:  "ip-data",
-		Data: ipLookupRes,
-	}, logger.LoggerOptions{
-		Key:  "user-agent",
-		Data: *agent,
-	})
-
-	ctx.SetContextData("Latitude", ipLookupRes.Latitude)
-	ctx.SetContextData("Longitude", ipLookupRes.Longitude)
-
-	ctx.UserAgent = agent
-	ctx.DeviceName = utils.GetStringPointer(fmt.Sprintf("%s/%s", agentDetails.Device, agentDetails.Name))
-	ctx.DeviceID = ctx.GetHeader("X-Device-Id")
-	if ctx.DeviceID == nil || *ctx.DeviceID == "" {
+	ctx.UserAgent = *agent
+	ctx.DeviceName = fmt.Sprintf("%s/%s", agentDetails.Device, agentDetails.Name)
+	deviceID := ctx.GetHeader("X-Device-Id")
+	if deviceID == nil || *deviceID == "" {
 		apperrors.MalformedHeader(ctx.Ctx)
 		return nil, false
 	}
+	ctx.DeviceID = *deviceID
 	return ctx, true
 }
